requirements.txt
streamlit
google-generativeai
PyGithub